Move base environment variables to executionspace creation
This will make it easier to create executionspace providers, where
the custom provider will just add environment variables it needs
added or want to override instead of knowing all variables that
are required for ETOS.
