+++ pkg/provider/executionspace.go
@@ -17,10 +17,14 @@ package provider
 
 import (
 	"context"
+	"errors"
 	"fmt"
+	"maps"
+	"os"
 
 	"github.com/eiffel-community/etos/api/v1alpha1"
 	"github.com/eiffel-community/etos/api/v1alpha2"
+	"github.com/fernet/fernet-go"
 	"github.com/go-logr/logr"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
@@ -91,6 +95,17 @@ func CreateExecutionSpace(
 	if err != nil {
 		return nil, err
 	}
+	environmentVariables, err := environmentVariables(ctx, environmentrequest)
+	if err != nil {
+		return nil, err
+	}
+
+	// Copies environment variables from the spec into the environmentVariables.
+	// This means that spec.Instructions.Environment will overwrite any environment
+	// variables with the same name as those generated from the EnvironmentRequest config.
+	maps.Copy(environmentVariables, spec.Instructions.Environment)
+	spec.Instructions.Environment = environmentVariables
+
 	labels := map[string]string{
 		"app.kubernetes.io/name":    "execution-space-provider",
 		"app.kubernetes.io/part-of": "etos",
@@ -136,4 +151,86 @@ func DeleteExecutionSpace(ctx context.Context, executionSpace *v1alpha2.Executio
 		return err
 	}
 	return cli.Delete(ctx, executionSpace)
+}
+
+// environmentVariables gets the environment variables for the Environment from the EnvironmentRequest
+// and encrypts sensitive information using the provided encryption key.
+func environmentVariables(
+	ctx context.Context,
+	environmentrequest *v1alpha1.EnvironmentRequest,
+) (map[string]string, error) {
+	var environment map[string]string
+	cli, err := KubernetesClient()
+	if err != nil {
+		return environment, err
+	}
+	key, err := environmentrequest.Spec.Config.EncryptionKey.Get(ctx, cli, environmentrequest.Namespace)
+	if err != nil {
+		return environment, errors.Join(errors.New("failed to get encryption key"), err)
+	}
+	encryptionKey, err := fernet.DecodeKey(string(key))
+	if err != nil {
+		return environment, errors.Join(errors.New("failed to decode encryption key"), err)
+	}
+	hostname, err := os.Hostname()
+	if err != nil {
+		return environment, errors.Join(errors.New("failed to get hostname"), err)
+	}
+	etosMessagebusPassword, err := getAndEncrypt(ctx,
+		cli, environmentrequest.Spec.Config.EtosMessageBus.Password,
+		environmentrequest.Namespace, encryptionKey,
+	)
+	if err != nil {
+		return environment, errors.Join(errors.New("failed to get and encrypt ETOS MessageBus password"), err)
+	}
+	eiffelMessagebusPassword, err := getAndEncrypt(ctx,
+		cli, environmentrequest.Spec.Config.EiffelMessageBus.Password,
+		environmentrequest.Namespace, encryptionKey,
+	)
+	if err != nil {
+		return environment, errors.Join(errors.New("failed to get and encrypt Eiffel MessageBus password"), err)
+	}
+
+	environment = map[string]string{
+		"SOURCE_HOST":                 hostname,
+		"ETOS_API":                    environmentrequest.Spec.Config.EtosApi,
+		"ETR_VERSION":                 environmentrequest.Spec.Providers.ExecutionSpace.TestRunner,
+		"ETOS_GRAPHQL_SERVER":         environmentrequest.Spec.Config.GraphQlServer,
+		"ETOS_RABBITMQ_EXCHANGE":      environmentrequest.Spec.Config.EtosMessageBus.Exchange,
+		"ETOS_RABBITMQ_HOST":          environmentrequest.Spec.Config.EtosMessageBus.Host,
+		"ETOS_RABBITMQ_PASSWORD":      string(etosMessagebusPassword),
+		"ETOS_RABBITMQ_PORT":          environmentrequest.Spec.Config.EtosMessageBus.Port,
+		"ETOS_RABBITMQ_USERNAME":      environmentrequest.Spec.Config.EtosMessageBus.Username,
+		"ETOS_RABBITMQ_VHOST":         environmentrequest.Spec.Config.EtosMessageBus.Vhost,
+		"ETOS_RABBITMQ_SSL":           environmentrequest.Spec.Config.EtosMessageBus.SSL,
+		"RABBITMQ_EXCHANGE":           environmentrequest.Spec.Config.EiffelMessageBus.Exchange,
+		"RABBITMQ_HOST":               environmentrequest.Spec.Config.EiffelMessageBus.Host,
+		"RABBITMQ_PASSWORD":           string(eiffelMessagebusPassword),
+		"RABBITMQ_PORT":               environmentrequest.Spec.Config.EiffelMessageBus.Port,
+		"RABBITMQ_USERNAME":           environmentrequest.Spec.Config.EiffelMessageBus.Username,
+		"RABBITMQ_VHOST":              environmentrequest.Spec.Config.EiffelMessageBus.Vhost,
+		"RABBITMQ_SSL":                environmentrequest.Spec.Config.EiffelMessageBus.SSL,
+		"OTEL_EXPORTER_OTLP_ENDPOINT": os.Getenv("OTEL_EXPORTER_OTLP_ENDPOINT"),
+		"OTEL_EXPORTER_OTLP_INSECURE": os.Getenv("OTEL_EXPORTER_OTLP_INSECURE"),
+	}
+	return environment, nil
+}
+
+// encrypt encrypts a string using the provided Fernet key.
+func encrypt(s []byte, key *fernet.Key) ([]byte, error) {
+	return fernet.EncryptAndSign(s, key)
+}
+
+// getAndEncrypt gets a value from a Var struct and encrypts it using the provided Fernet key.
+func getAndEncrypt(
+	ctx context.Context, cli client.Client, s *v1alpha1.Var, namespace string, key *fernet.Key,
+) ([]byte, error) {
+	if s == nil {
+		return nil, errors.New("no value provided")
+	}
+	value, err := s.Get(ctx, cli, namespace)
+	if err != nil {
+		return nil, err
+	}
+	return encrypt(value, key)
 }
